Add framebuffer
Framebuffer is set up using the tag mailbox. Screen size is determined by
asking the VC for the current screen size, then setting up the screen with
that size. Not sure if this is the correct way to do it, but it seems to
work.

Character set is the SAA5050 teletext set, familiar to BBC Micro users.