Bug fix: screen sizes other than 1280x720
The console was assuming the framebuffer was 1280 pixels wide (which it is
on my Pi), rather than using the actual value, so text will look very wrong
on other monitor sizes.

Text should now display properly on other screen sizes, rotated screens,
etc.

The way the address for a pixel is calculated is still wrong, and won't work
if the screen isn't a multiple of 8 pixels wide. It should be using the
pitch value from VideoCore, not the width of the framebuffer, to calculate
the vertical address offset. This is probably related to the alignment value
when requesting a framebuffer allocation.

Also, minor addition to README