Include libgcc.a
Where it can't put code inline to perform a particular task (eg, division),
gcc will place a function call to a function to do the task instead. These
functions are in libgcc.a

Unfortunately, while the Raspbian gcc has a libgcc.a suitable for a
Raspberry Pi, some other compilers don't - eg. my cross compiler on x86 has
a libgcc containing Thumb2 instructions. This leads to undefined instruction
exceptions and hilarity*

Included a copy of libgcc.a taken from the Raspbian "Wheezy" distribution,
for ease of building. The makefile will detect if it's being run on a
Raspberry Pi, and use rpi-libgcc/libgcc.a if not.

Also included the licence files for libgcc.a, and a "licence" for this code.

divby0.c contains a routine which libgcc.a's divide function expects to be
able to call if an attempt is made to divide by 0. For now, it prints an
error and stops the kernel.

Included -gc-sections in the linker options to remove unnecessary code (such
as the function in divby0.c above if gcc hasn't placed a library call to do
division)

Update the README to include all of this

*not hilarious. Vexing.