Add interrupt handling
Add code to handle interrupts, including IRQ, SWI and abort conditions (not
yet used).

The interrupt code sets up a set of interrupt vectors of 0x00000000, then
enables interrupts for the ARM timer and uses that to flash the OK LED.
There is also a test SWI. Stacks have been added for all processor modes,
though most will never see more than a handful of bytes in them.

Removed the code which makes the LED reflect the sate of GPIO14. It was
never the best of ideas.